requirements.txt
# Core Dependencies
streamlit>=1.28.0
python-dotenv>=1.0.0

# IBM Watsonx and LangChain
langchain-ibm>=0.1.0
langchain-core>=0.1.0
langgraph>=0.0.40

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.0.0

# Database (SQLite only - built into Python)
# Note: sqlite3 is included with Python by default
# Optional vector extension (if available)
# sqlite-vec>=0.1.0

# Additional Utilities
typing-extensions>=4.7.0
plotly>=5.0.0